Use custom component for stations
- Pull station code out of WisMap component into WisStation.
- More informative hover when station has not been linked to collection.
